+++ src/test/scala/edu/rit/cs/mmior/jsonoid/discovery/schemas/ProductSchemaSpec.scala
@@ -22,7 +22,7 @@ class ProductSchemaSpec extends UnitSpec {
   private val allProps = SchemaProperties
     .empty[JsonSchema[_]]
     .replaceProperty(
-      ProductSchemaTypesProperty(List(StringSchema()), List(1), true)
+      ProductSchemaTypesProperty(AnySchema(), List(StringSchema()), List(1), true)
     )
   private val allProductSchema = ProductSchema(allProps)